Add a Home screen that will be the first screen if already onboarded, or the screen after onboarding.
Added a profile image on the top of the navigation bar for the Home and Profile screens. On the Home screen, this can be selected to open the Profile screen.
For the Home screen, obtain the menu list and info from a URL at first run, store the data in an SQLite database, and load from that on future runs.
For the Profile screen, make sure that the profile image on the navigation bar is updated when the image is selected or removed.
For the menu list on the Home screen, provide category selection buttons that can be independently selected to indicate which menu categories should be include (all are included if no categories are selected), as well as a text search field to apply to names of the menu items to be included.